Changed the performance test suite to support mulitple configurations
The suite was previously only testing in-memory configurations (e.g., no persistence), which isn't really
a realistic scenario. (It is still useful and interesting as a test.) So I changed the framework to allow
multiple configurations, where each configuration is set via a Maven profile. Currently there are two
configurations: "local-inmemory" and "local-filesystem".